fix(sln): repair PrimeTime filter; split mock templates partial
- Point Nexo.PrimeTime.slnf at application test projects; add Nexo.Tests.CLI to Nexo.sln
- Split MockScaffoldingResponder into partial Core + Templates files (behavior unchanged)
- Clarify Nexo.Compat/ in ProjectTiers as source-only polyfills
