+++ Nexo.PrimeTime.slnf
@@ -4,9 +4,9 @@
     "projects": [
       "src/Nexo.Tests.Application/Nexo.Tests.Application.csproj",
       "src/Nexo.Tests.BackgroundAgents/Nexo.Tests.BackgroundAgents.csproj",
-      "src/Nexo.Tests.CLI/Nexo.Tests.CLI.csproj",
+      "application/src/Nexo.Tests.CLI/Nexo.Tests.CLI.csproj",
       "src/Nexo.Tests.Domain/Nexo.Tests.Domain.csproj",
-      "src/Nexo.Tests.GameDomain/Nexo.Tests.GameDomain.csproj",
+      "application/src/Nexo.Tests.GameDomain/Nexo.Tests.GameDomain.csproj",
       "src/Nexo.Tests.Infrastructure/Nexo.Tests.Infrastructure.csproj",
       "src/Nexo.Tests.Orchestration/Nexo.Tests.Orchestration.csproj",
       "src/Nexo.Tests.Transport/Nexo.Tests.Transport.csproj"
