+++ Nexo.sln
@@ -105,6 +105,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nexo.Tests.GameDomain", "ap
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nexo.Infrastructure", "src\Nexo.Infrastructure\Nexo.Infrastructure.csproj", "{FB24C010-F1E8-4C0D-AB0C-58CFC74A840C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nexo.Tests.CLI", "application\src\Nexo.Tests.CLI\Nexo.Tests.CLI.csproj", "{71D3B105-0223-4236-AB5A-A96D30604642}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -691,6 +693,18 @@ Global
 		{FB24C010-F1E8-4C0D-AB0C-58CFC74A840C}.Release|x64.Build.0 = Release|Any CPU
 		{FB24C010-F1E8-4C0D-AB0C-58CFC74A840C}.Release|x86.ActiveCfg = Release|Any CPU
 		{FB24C010-F1E8-4C0D-AB0C-58CFC74A840C}.Release|x86.Build.0 = Release|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Debug|x64.Build.0 = Debug|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Debug|x86.Build.0 = Debug|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Release|Any CPU.Build.0 = Release|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Release|x64.ActiveCfg = Release|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Release|x64.Build.0 = Release|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Release|x86.ActiveCfg = Release|Any CPU
+		{71D3B105-0223-4236-AB5A-A96D30604642}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -744,6 +758,7 @@ Global
 		{9CD88C08-6A2D-44C2-85BB-0E1A45D08383} = {F89E166B-875E-4F45-A72F-DC15564E7E9A}
 		{1034F18B-417F-45EF-A736-A7DC3250D3B0} = {F89E166B-875E-4F45-A72F-DC15564E7E9A}
 		{FB24C010-F1E8-4C0D-AB0C-58CFC74A840C} = {9D4F8B1A-0B6E-4A3E-8A6A-0DE12C7C6E2F}
+		{71D3B105-0223-4236-AB5A-A96D30604642} = {F89E166B-875E-4F45-A72F-DC15564E7E9A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {12345678-1234-1234-1234-123456789ABC}
